requirements-latest.txt
numpy==2.3.2
pandas==2.3.1
oct2py==5.8.0
matplotlib==3.10.5
openpyxl==3.1.5

matpowercaseframes==1.1.4
pypglib==0.0.3

pytest==8.4.1
pytest-cov==6.2.1
pytest-xdist==3.8.0

jupyter==1.1.1
nbmake==1.5.5

setuptools==80.9.0

pre-commit==3.7.1
nb-clean==4.0.1
ruff==0.12.9
